refactor: user-related types and schemas are now in user.ts file.
